temp fix for new "signal emitted with wrong data type" errors
This is a temporary fix, because signal values will need to be revised anyway

+++ src/inet/linklayer/mrp/Mrp.cc
@@ -1120,7 +1120,7 @@ void Mrp::setupTopologyChangeReq(simtime_t interval) {
     packet2->insertAtBack(endTlv);
     MacAddress sourceAddress2 = getPortNetworkInterface(secondaryRingPort)->getMacAddress();
     sendFrameReq(secondaryRingPort, static_cast<MacAddress>(MC_CONTROL), sourceAddress2, priority, MRP_LT, packet2);
-    emit(topologyChangeSignal, interval.inUnit(SIMTIME_MS));
+    emit(topologyChangeSignal, (uint64_t)interval.inUnit(SIMTIME_MS));
 }
 
 void Mrp::setupLinkChangeReq(int ringPort, uint16_t linkState, simtime_t time) {
@@ -1154,7 +1154,7 @@ void Mrp::setupLinkChangeReq(int ringPort, uint16_t linkState, simtime_t time) {
     packet1->insertAtBack(endTlv);
     MacAddress sourceAddress1 = getPortNetworkInterface(ringPort)->getMacAddress();
     sendFrameReq(ringPort, static_cast<MacAddress>(MC_CONTROL), sourceAddress1, priority, MRP_LT, packet1);
-    emit(linkChangeSignal, time.inUnit(SIMTIME_MS));
+    emit(linkChangeSignal, (double)time.inUnit(SIMTIME_MS));
 }
 
 void Mrp::testMgrNackReq(int ringPort, MrpPriority managerPrio, MacAddress sourceAddress) {
